refactor: extract earnings and stock-action tools into market/earnings.py (#576)
