+++ tests/unit/test_market_data_module_split.py
@@ -1,5 +1,14 @@
 """Identity assertions for market data module split compatibility."""
 
+from open_stocks_mcp.tools.market.earnings import (
+    get_stock_earnings as earnings_get_stock_earnings,
+)
+from open_stocks_mcp.tools.market.earnings import (
+    get_stock_events as earnings_get_stock_events,
+)
+from open_stocks_mcp.tools.market.earnings import (
+    get_stock_splits as earnings_get_stock_splits,
+)
 from open_stocks_mcp.tools.market.movers import (
     get_stocks_by_tag as movers_get_stocks_by_tag,
 )
@@ -23,6 +32,15 @@ from open_stocks_mcp.tools.robinhood_market_data_tools import (
 )
 from open_stocks_mcp.tools.robinhood_market_data_tools import (
     get_stock_ratings as legacy_get_stock_ratings,
+)
+from open_stocks_mcp.tools.robinhood_market_data_tools import (
+    get_stock_earnings as legacy_get_stock_earnings,
+)
+from open_stocks_mcp.tools.robinhood_market_data_tools import (
+    get_stock_events as legacy_get_stock_events,
+)
+from open_stocks_mcp.tools.robinhood_market_data_tools import (
+    get_stock_splits as legacy_get_stock_splits,
 )
 from open_stocks_mcp.tools.robinhood_market_data_tools import (
     get_stocks_by_tag as legacy_get_stocks_by_tag,
@@ -58,3 +76,12 @@ class TestMarketDataModuleSplit:
 
     def test_get_stock_news_identity(self) -> None:
         assert legacy_get_stock_news is news_get_stock_news
+
+    def test_get_stock_earnings_identity(self) -> None:
+        assert legacy_get_stock_earnings is earnings_get_stock_earnings
+
+    def test_get_stock_splits_identity(self) -> None:
+        assert legacy_get_stock_splits is earnings_get_stock_splits
+
+    def test_get_stock_events_identity(self) -> None:
+        assert legacy_get_stock_events is earnings_get_stock_events
