+++ tests/unit/test_market_research_tools.py
@@ -5,14 +5,16 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
+from open_stocks_mcp.tools.market.earnings import (
+    get_stock_earnings,
+    get_stock_events,
+    get_stock_splits,
+)
 from open_stocks_mcp.tools.market.movers import get_top_movers_sp500
 from open_stocks_mcp.tools.market.news import get_stock_news
 from open_stocks_mcp.tools.market.ratings import get_stock_ratings
 from open_stocks_mcp.tools.robinhood_market_data_tools import (
-    get_stock_earnings,
-    get_stock_events,
     get_stock_level2_data,
-    get_stock_splits,
 )
 
 
@@ -294,9 +296,9 @@ class TestStockEarnings:
 
     @pytest.mark.journey_research
     @pytest.mark.unit
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.earnings.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.earnings.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.earnings.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
@@ -367,9 +369,9 @@ class TestStockEarnings:
     @pytest.mark.skip(reason="Slow exception test - run with pytest -m exception_test")
     @pytest.mark.journey_research
     @pytest.mark.unit
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.earnings.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.earnings.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.earnings.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
@@ -502,9 +504,9 @@ class TestStockNews:
 class TestStockSplits:
     """Test stock splits functionality."""
 
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.earnings.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.earnings.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.earnings.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
@@ -565,9 +567,9 @@ class TestStockSplits:
 
     @pytest.mark.exception_test
     @pytest.mark.skip(reason="Slow exception test - run with pytest -m exception_test")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.earnings.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.earnings.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.earnings.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
@@ -597,9 +599,9 @@ class TestStockSplits:
 class TestStockEvents:
     """Test stock events functionality."""
 
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.earnings.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.earnings.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.earnings.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
@@ -656,9 +658,9 @@ class TestStockEvents:
 
     @pytest.mark.exception_test
     @pytest.mark.skip(reason="Slow exception test - run with pytest -m exception_test")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.earnings.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.earnings.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.earnings.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
